Potential Evapotranspiration

This function calculates Penman-Monteith, Priestley Taylor and Hargreaves-Samani Potential Evapotranspiration using the method described by Allen et al, (1998)

Usage

calcEto(data, method = "PM", crop = "short", Zh = NULL)

Arguments

data

A dataframe containing the required weather variables with the following columns:

Lat: Latitude of the site in decimal degrees.

Lon: Longitude of the site in decimal degrees.

Elev: Elevation above sea level in meters.

Year: Year of record "YYYY".

Month: Month of record "MM".

Day: Day of record "DD".

Tmax: Daily maximum temperature at 2-m height in °C.

Tmin: Daily minimum temperature at 2-m height in °C.

Rs: Daily surface incoming solar radiation in MJ/m^2/day.

RH or RHmax and RHmin: Daily relative humidity at 2-m height.

Tdew: Daily dew point temperature at 2-m height in °C.

U2 or Uz: Daily wind speed at 2-m or custom height (m/s).

method

The formulation used to compute Eto; default is "PM" for Penman-Monteith, "PT" for Priestley-Taylor, and "HS" for Hargreaves-Samani.

crop

Either "short" (default) for FAO-56 hypothetical short grass or "tall" for ASCE-EWRI standard crop.

Zh

Height of wind speed measurement in meters.

Value

A list containing:

ET.Daily: Daily estimations of reference crop evapotranspiration (mm/day).

Ra.Daily: Daily estimations of extraterrestrial radiation (MJ/m^2/day).

Slope.Daily: Daily estimations of slope of vapor pressure curve (kPa/°C).

ET.type: Type of the estimation obtained.

Details

Penman-Monteith: If all variables of Tmax, Tmin, Rs, either U2 or Uz, and either RHmax and RHmin or RH or Tdew are available and crop surface (short or tall) is specified, the Penman-Monteith FAO56 formulation is used (Allen et al. 1998).

Priestley-Taylor: If all variables of Tmax, Tmin, Rs, and either RHmax and RHmin or RH or Tdew are available, the Priestley-Taylor formulation is used (Priestley and Taylor, 1972).

Hargreaves-Samani: If only Tmax and Tmin are available, the Hargreaves-Samani formulation is used for estimating reference crop evapotranspiration (Hargreaves and Samani, 1985).

References

Allen, R.G., L.S. Pereira, D. Raes, and M. Smith. 1998. Crop evapotranspiration-Guidelines for Computing Crop Water requirements FAO Irrigation and Drainage Paper 56. FAO, Rome 300: 6541.

Allen, R. G. 2005. The ASCE standardized reference evapotranspiration equation. Amer Society of Civil Engineers.

Guo, D., Westra, S., & Maier, H. (2016). An R package for modelling actual, potential and reference evapotranspiration. Environmental Modelling & Software, 78, 216-224. doi:10.1016/j.envsoft.2015.12.019.

Hargreaves, G.H., & Samani, Z.A. (1985). Reference crop evapotranspiration from ambient air temperature. American Society of Agricultural Engineers.

Priestley, C., & Taylor, R. (1972). On the assessment of surface heat flux and evaporation using large-scale parameters. Monthly Weather Review, 100(2), 81-92.

See also

Examples

## Load sample data:
data(climateData)
PET.HS <- calcEto(climateData, method = "HS")
#> Warning: Warning: There are 1562 missing values in the data:  Tmax
#> Warning: Warning: There are 847 missing values in the data:  Tmin
#> Warning: Warning: There are 2 instances where the maximum temperature is less than the minimum temperature.
#> Adjusted 2 instances where Tmin was equal to or greater than Tmax.

## Load sample data:
data(AgroClimateData)
PET.PM <- calcEto(AgroClimateData, method = "PM", crop = "short")
#> Adjusted 0 instances where Tmin was equal to or greater than Tmax.
#> Penman-Monteith FAO56 Reference Crop ET
#> Evaporative surface: FAO-56 hypothetical short grass, albedo = 0.23 ; surface resistance = 70 sm^-1; crop height = 0.12 m; roughness height = 0.02 m
#> Timestep: daily
#> Units: mm
#> Time duration: 1982-01-01 to 2022-12-31