Daily Weather Data (1996-2020) from Angochen Weather Observing Station, Mozambique

Contains daily weather data obtained from Instituto Nacional de Meteorologia (INAM). The sample data includes daily raw weather data from January 1, 1996, to December 31, 2020, from a weather station located in Angochen, Nampula province of Mozambique.

Usage

data(climateData)

Format

A data frame containing daily values of weather variables:

Station_Name: Name of the weather station.

Lat: Latitude of the site in decimal degrees.

Lon: Longitude of the site in decimal degrees.

Elev: Elevation above sea level in meters.

Year: Year of the record (YYYY).

Month: Month of the record (MM).

Day: Day of the record (DD).

Rain: Daily rainfall in millimeters (mm).

Tmax: Daily maximum temperature at 2 meters height in degrees Celsius (°C).

Tmin: Daily minimum temperature at 2 meters height in degrees Celsius (°C).

Source

INAM - Instituto Nacional de Meteorologia, Mozambique

See also

Examples

## Load the climate data
data(climateData)

## Get the structure of the data frame
str(climateData)
#> 'data.frame':	9132 obs. of  10 variables:
#>  $ Station_Name: chr  "ANGOCHE" "ANGOCHE" "ANGOCHE" "ANGOCHE" ...
#>  $ Lat         : num  -16.2 -16.2 -16.2 -16.2 -16.2 ...
#>  $ Lon         : num  39.9 39.9 39.9 39.9 39.9 ...
#>  $ Elev        : num  25.2 25.2 25.2 25.2 25.2 ...
#>  $ Year        : num  1996 1996 1996 1996 1996 ...
#>  $ Month       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Day         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Rain        : num  0 0 7 0 0 0 0 7 0 0 ...
#>  $ Tmax        : num  31.5 32 30 30.9 31.4 31.5 32 30 30.9 31.4 ...
#>  $ Tmin        : num  22.5 22 22.8 24.5 24 22.5 22 22.8 24.5 24 ...

## Get the head of the data frame
head(climateData)
#>   Station_Name      Lat     Lon  Elev Year Month Day Rain Tmax Tmin
#> 1      ANGOCHE -16.2163 39.9145 25.19 1996     1   1    0 31.5 22.5
#> 2      ANGOCHE -16.2163 39.9145 25.19 1996     1   2    0 32.0 22.0
#> 3      ANGOCHE -16.2163 39.9145 25.19 1996     1   3    7 30.0 22.8
#> 4      ANGOCHE -16.2163 39.9145 25.19 1996     1   4    0 30.9 24.5
#> 5      ANGOCHE -16.2163 39.9145 25.19 1996     1   5    0 31.4 24.0
#> 6      ANGOCHE -16.2163 39.9145 25.19 1996     1   6    0 31.5 22.5