Daily Surface Meteorological Data (1982-2022) Extracted from AgERA5

The AgERA5 dataset provides daily surface meteorological data for the period from 1979 to present as input for agriculture and agro-ecological studies. This dataset is based on the hourly ECMWF ERA5 data at surface level, with data from January 1, 1982, through December 31, 2022, extracted for a grid located in Angochen, Nampula province of Mozambique.

Usage

data(AgroClimateData)

Format

A data frame containing daily observations of weather variables:

GridID: Grid ID of the location.

Lat: Latitude of the grid in decimal degrees.

Lon: Longitude of the grid in decimal degrees.

Elev: Elevation above sea level in meters.

WHC: Water holding capacity in millimeters (mm).

Year: Year of record ("YYYY").

Month: Month of record ("MM").

Day: Day of record ("DD").

Rain: Daily precipitation in millimeters (mm/day).

Tmax: Daily maximum temperature at 2 meters height in degrees Celsius (°C).

Tmin: Daily minimum temperature at 2 meters height in degrees Celsius (°C).

Rs: Total amount of energy provided by solar radiation at the surface over the period 00-24h local time per unit area and time (MJ/m^2/day).

Tdew: Mean dewpoint temperature at a height of 2 metres above the surface over the period 00h-24h local time (°C).

Uz: Mean wind speed at a height of 2 metres above the surface over the period 00h-24h local time (m/s).

Source

https://cds.climate.copernicus.eu/cdsapp#!/dataset/sis-agrometeorological-indicators?tab=overview

References

AgERA5, 2021, Copernicus Climate Change Service (C3S), Fifth generation of ECMWF atmospheric reanalysis of the global climate for agriculture and agro-ecological studies. Copernicus Climate Change Service Climate Data Store (CDS), July 2021.

See also

Examples

## Load the agroclimate data
data(AgroClimateData)

## Get the structure of the data frame
str(AgroClimateData)
#> 'data.frame':	14975 obs. of  14 variables:
#>  $ GridID: chr  "MOZ0007149" "MOZ0007149" "MOZ0007149" "MOZ0007149" ...
#>  $ Lat   : num  -15.1 -15.1 -15.1 -15.1 -15.1 ...
#>  $ Lon   : num  39.3 39.3 39.3 39.3 39.3 ...
#>  $ Elev  : num  392 392 392 392 392 ...
#>  $ WHC   : num  97.8 97.8 97.8 97.8 97.8 ...
#>  $ Year  : num  1982 1982 1982 1982 1982 ...
#>  $ Month : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Day   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Rain  : num  0 0 0 1.91 0 ...
#>  $ Tmax  : num  32.2 33.1 33.5 32.8 32.7 ...
#>  $ Tmin  : num  23.1 23.1 23.1 23.6 22.8 ...
#>  $ Rs    : num  23.9 26.4 25 24.2 23.4 ...
#>  $ Tdew  : num  20.2 20.5 20.5 20.8 21.4 ...
#>  $ Uz    : num  4.72 4.28 3.62 2.54 1.48 ...

## Get the head of the data frame
head(AgroClimateData)
#>       GridID       Lat     Lon     Elev      WHC Year Month Day     Rain
#> 1 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   1 0.000000
#> 2 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   2 0.000000
#> 3 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   3 0.000000
#> 4 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   4 1.907393
#> 5 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   5 0.000000
#> 6 MOZ0007149 -15.09238 39.2519 392.1337 97.84914 1982     1   6 0.000000
#>       Tmax     Tmin       Rs     Tdew       Uz
#> 1 32.24396 23.11500 23.86698 20.21160 4.723783
#> 2 33.07202 23.12585 26.38375 20.48284 4.279407
#> 3 33.49679 23.12602 25.00704 20.45689 3.622179
#> 4 32.76818 23.60351 24.16475 20.83896 2.535047
#> 5 32.65872 22.79294 23.44483 21.36882 1.477617
#> 6 31.80630 22.43975 21.99277 21.29297 1.953415