Rainy Season Calendar

This function estimates the rainy season calendar, i.e onset date, cessation date and duration of the rainy season based on Agroclimatic approach. The agroclimatic approach defines the onset of the rainy season as the optimal date that ensures sufficient soil moisture during planting and early growing periods to avoid crop failure after sowing and requires information on rainfall, reference evapotranspiration and accounting of the daily soil water balance parameters.

Usage

calcSeasCal(data, onsetWind.start, onsetWind.end, cessaWind.end, soilWHC)

Arguments

data

an R object of dataframe as returned by calcWatBal or a dataframe having similar parameters.

onsetWind.start

Earliest possible start date of the onset window.

onsetWind.end

The latest possible date for end of the onset window.

cessaWind.end

The latest possible date for end of the cessation window.

soilWHC

Water holding capacity of the soil at root zone depth in (mm).

Value

The function generates list of dataframes with columns of variables:

Onset.dF: a data frame with columns of onset variables:

Year: year of the season under investigation, "YYYY".

onset.Year: year of the season under investigation, "YYYY".

onset.Month: month of the onset of the season in "MM".

onset.Day: day of the onset of the season in "DD".

JD: onset date of the season in Julian day, "DOY".

YYYYDOY: onset date of the season in "YYYY-DOY".

Year: year of the season under investigation, "YYYY".

Cessation.dF: a data frame with columns of onset variables:

Year: year of the season under investigation, "YYYY".

cessation.Year: year of the season under investigation, "YYYY".

cessation.Month: month of the cessation of the season in "MM".

cessation.Day: day of the cessation of the season in "DD".

JD: cessation date of the season in Julian day, "DOY".

YYYYDOY: cessation date of the season in "YYYY-DOY".

Duration.dF: a data frame with columns of onset variables:

Year: year of the season under investigation, "YYYY".

onset.YYYYDOY: onset date of the season in "YYYY-DOY".

cessation.YYYYDOY: cessation date of the season in "YYYY-DOY".

Duration: duration of the season in "days".

Details

As per agroclimatic approach, a normal rainy season (growing season) is defined as one when there is an excess of precipitation over potential evapotranspiration (PET). Such a period met the evapotransiration demands of crops and recharge the moisture of the soil profile (FAO 1977; 1978; 1986). Thus, the rainy season calendar defined accordingly:

Onset

The onset of the rainy season will start on the first day after onsetWind.start, when the actual-to-potential evapotranspiration ratio is greater than 0.5 for 7 consecutive days, followed by a 20-day period in which plant available water remains above wilting over the root zone of the soil layer.

Cesation

The rainy season will end, cessation, on the first day after onsetWind.end, when the actual-to-potential evapotranspiration ratio is less than 0.5 for 7 consecutive days, followed by 12 consecutive non-growing days in which plant available water remains below wilting over the root zone of the soil layer.

Duration

The duration of the rainy season is the total number of days from onset to cessation of the season.

References

FAO, 1977. Crop water requirements. FAO Irrigation and Drainage Paper No. 24, by Doorenbos J and W.O. Pruitt. FAO, Rome, Italy.

FAO 1978. Forestry for Local Community Development Food and Agriculture Organization of the United Nation (FAO), FAO Forestry paper, No 7, Rome.

FAO, 1986. Early Agrometeorological crop yield forecasting. FAO Plant Production and Protection paper No. 73, by M. Frère and G.F. Popov. FAO, Rome, Italy

See also

Examples

# load example data:

data(AgroClimateData)

# Estimate daily PET:
PET <- calcEto(AgroClimateData, method = "PM", Zh = 10)

# Add the estimated PET 'ET.Daily' to a new column in AgroClimateData:
AgroClimateData$Eto <- PET$ET.Daily

# Estimate daily water balance for the soil having 100mm of WHC:
watBal <- calcWatBal(AgroClimateData, soilWHC = 100)

# estimate the rainy season calandar (Onset, Cessation and Duration):
onsetWind.start = "1996-09-01"  # earliest possible start data of the onset window
onsetWind.end = "1997-01-31"   # the late possible date for end of the onset window
cessaWind.end = "1997-06-30"   # the late possible date for end of the cessation window

seasCal.lst <- calcSeasCal(watBal, onsetWind.start, onsetWind.end, cessaWind.end, soilWHC = 100)

str(seasCal.lst)
#> List of 3
#>  $ :'data.frame':	40 obs. of  6 variables:
#>   ..$ Year       : num [1:40] 1982 1983 1984 1985 1986 ...
#>   ..$ onset.Year : num [1:40] 1982 1983 1984 1985 1986 ...
#>   ..$ onset.Month: num [1:40] 11 12 11 11 11 1 12 11 12 12 ...
#>   ..$ onset.Day  : num [1:40] 11 11 9 19 20 10 18 13 26 3 ...
#>   ..$ onset.JD   : num [1:40] 315 345 314 323 324 10 353 317 360 337 ...
#>   ..$ onset.Value: int [1:40] 72 102 70 80 81 132 109 74 117 94 ...
#>  $ :'data.frame':	40 obs. of  6 variables:
#>   ..$ Year           : num [1:40] 1982 1983 1984 1985 1986 ...
#>   ..$ cessation.Year : num [1:40] 1983 1984 1985 1986 1987 ...
#>   ..$ cessation.Month: num [1:40] 5 4 5 5 4 4 5 3 5 5 ...
#>   ..$ cessation.Day  : num [1:40] 12 18 9 10 22 21 6 15 6 21 ...
#>   ..$ cessation.JD   : num [1:40] 132 109 129 130 112 112 126 74 126 142 ...
#>   ..$ cessation.Value: num [1:40] 255 232 252 253 235 235 249 197 249 265 ...
#>  $ :'data.frame':	40 obs. of  4 variables:
#>   ..$ Year             : num [1:40] 1982 1983 1984 1985 1986 ...
#>   ..$ onset.YYYYDOY    : chr [1:40] "1982-315" "1983-345" "1984-314" "1985-323" ...
#>   ..$ cessation.YYYYDOY: chr [1:40] "1983-132" "1984-109" "1985-129" "1986-130" ...
#>   ..$ Duration         : int [1:40] 183 130 182 173 154 103 140 123 132 171 ...