Daily Soil Water Balance Estimation

This function estimates the daily soil water balance based on a simple mass balance budget approach. It calculates the amount of water available in the root zone of a homogeneous grass canopy growing on a well-drained, homogeneous soil.

Usage

calcWatBal(data, soilWHC)

Arguments

data

A data frame containing the required input variables. The data frame must include the following columns:

Lat: Latitude of the site (decimal degrees).

Lon: Longitude of the site (decimal degrees).

Elev: Elevation above sea level (meters).

Year: Year of the record ("YYYY").

Month: Month of the record ("MM").

Day: Day of the record ("DD").

Rain: Daily rainfall (mm).

Eto: Daily potential evapotranspiration (mm).

soilWHC

Numeric. Water holding capacity of the soil (mm).

Value

A list with the following components:

data

A data frame containing the results of the water balance calculations, with the following columns:

DRAIN: Deep drainage (mm).

TRAN: Water lost by transpiration (mm).

RUNOFF: Surface runoff (mm).

AVAIL: Available soil moisture storage (mm).

R: Ratio of actual to potential evapotranspiration.

warnings

A list of warnings related to any unrealistic or adjusted values in the input data or parameters used during the water balance calculations.

References

Allen, R.G.; Pereira, L.S.; Raes, D.; Smith, M. (1998). Crop Evapotranspiration: Guidelines for Computing Crop Water Requirements. FAO Irrigation and Drainage Paper no. 56, FAO: Rome, Italy. ISBN 92-5-104219-5.

Doorenbos, J., & Pruitt, W.O. (1975). Guidelines for Predicting Crop Water Requirements. Irrigation and Drainage Paper 24, Food and Agriculture Organization of the United Nations, Rome, 179 p.

See also

Examples

# \donttest{
## Load sample data
data(AgroClimateData)

# Estimate daily PET using the Penman-Monteith method
PET.PM <- calcEto(AgroClimateData, method = "PM", Zh = 10)
#> Adjusted 0 instances where Tmin was equal to or greater than Tmax.
#> Penman-Monteith FAO56 Reference Crop ET
#> Evaporative surface: FAO-56 hypothetical short grass, albedo = 0.23 ; surface resistance = 70 sm^-1; crop height = 0.12 m; roughness height = 0.02 m
#> Timestep: daily
#> Units: mm
#> Time duration: 1982-01-01 to 2022-12-31

# Add the estimated PET to the AgroClimateData frame
AgroClimateData$Eto <- PET.PM$ET.Daily

# Estimate daily soil water balance for a soil with 100 mm WHC
watBal <- calcWatBal(data = AgroClimateData, soilWHC = 100)
# }