A dataframe containing raw climate data

The R data object was obtained from Instituto Nacional de Meteorologia (INAM). This example data set contains the daily raw climate data over the period between 1/1/1996 and 12/31/2020 from a weather station located in Angochen, Nampula province of Mozambique.

Usage

data(climateData)

Format

A data frame containing daily observations of climate parameters:

Station_Name: name of the weather station.

Lat: latitude of the site in decimal degrees.

Lon: longitude of the site in decimal degrees.

Elev: elevation above sea level in (meters).

Year: year of record "YYYY".

Month: month of record "MM".

Day: day of record "DD".

Rain: daily rainfall in (mm).

Tmax: daily maximum temperature at 2-m height in (°C).

Tmin: daily minimum temperature at 2-m height in (°C).

Source

INAM - Instituto Nacional de Meteorologia, Mozambique

See also

Examples

# load example data:
data(climateData)

# Get the structure of the data frame:
str(climateData)
#> 'data.frame':	9132 obs. of  10 variables:
#>  $ Station_Name: chr  "ANGOCHE" "ANGOCHE" "ANGOCHE" "ANGOCHE" ...
#>  $ Lat         : num  -16.2 -16.2 -16.2 -16.2 -16.2 ...
#>  $ Lon         : num  39.9 39.9 39.9 39.9 39.9 ...
#>  $ Elev        : num  25.2 25.2 25.2 25.2 25.2 ...
#>  $ Year        : num  1996 1996 1996 1996 1996 ...
#>  $ Month       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Day         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Rain        : num  0 0 7 0 0 0 0 7 0 0 ...
#>  $ Tmax        : num  31.5 32 30 30.9 31.4 31.5 32 30 30.9 31.4 ...
#>  $ Tmin        : num  22.5 22 22.8 24.5 24 22.5 22 22.8 24.5 24 ...

# Get the head of the data frame:
head(climateData)
#>   Station_Name      Lat     Lon  Elev Year Month Day Rain Tmax Tmin
#> 1      ANGOCHE -16.2163 39.9145 25.19 1996     1   1    0 31.5 22.5
#> 2      ANGOCHE -16.2163 39.9145 25.19 1996     1   2    0 32.0 22.0
#> 3      ANGOCHE -16.2163 39.9145 25.19 1996     1   3    7 30.0 22.8
#> 4      ANGOCHE -16.2163 39.9145 25.19 1996     1   4    0 30.9 24.5
#> 5      ANGOCHE -16.2163 39.9145 25.19 1996     1   5    0 31.4 24.0
#> 6      ANGOCHE -16.2163 39.9145 25.19 1996     1   6    0 31.5 22.5